feat: offer backup downloads while a paused project is still restorable
